+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Version v0.5.1
 - Fixed a compatibility issue with numpy 2.0.0 (np.Inf was removed)
 - Updated errors due to missing dependencies
 - Basemap replaced with cartopy
+- Subgrid from shapefile now allows to use all features from a .shp file
 
 Version v0.5.0
 ==============